--- a/Protobuild.toml
+++ b/Protobuild.toml
@@ -23,4 +23,3 @@ generators = ["go"]
 # enable ttrpc and disable fieldpath and grpc for the shim
 prefixes = ["github.com/containerd/ttrpc/integration/streaming"]
 generators = ["go", "go-ttrpc"]
-plugins = ["ttrpc"]