@dmcgowan 2022-03-15
Marking this as ready for review

Benchmarking not showing a significant difference (only comparing unary requests obviously). There is one extra allocation in the new code, likely related to the new stream structure.
```
name                        old time/op    new time/op    delta
RoundTripUnixSocketCreds-8    26.3µs ± 0%    27.0µs ± 0%   ~     (p=1.000 n=1+1)
RoundTrip-8                   27.2µs ± 0%    25.5µs ± 0%   ~     (p=1.000 n=1+1)

name                        old alloc/op   new alloc/op   delta
RoundTripUnixSocketCreds-8    1.18kB ± 0%    1.31kB ± 0%   ~     (p=1.000 n=1+1)
RoundTrip-8                   1.18kB ± 0%    1.34kB ± 0%   ~     (p=1.000 n=1+1)

name                        old allocs/op  new allocs/op  delta
RoundTripUnixSocketCreds-8      27.0 ± 0%      28.0 ± 0%   ~     (p=1.000 n=1+1)
RoundTrip-8                     27.0 ± 0%      28.0 ± 0%   ~     (p=1.000 n=1+1)
```